split packets and charts tabs
